requirements.txt
Django==5.0
djangorestframework==3.15.2
psycopg2==2.9.9
python-dotenv
djangorestframework-simplejwt==5.3.1
celery==5.4.0
redis==5.0.7
flower==2.0.1
